chore(scheduler): split phase_continue_prefills 716 → 276 LoC (file-size-cap)
phase_continue_prefills.rs grew to 716 LoC after the Q12 batched-prefill
work (commits a542463, 2ff926d, 42612e6, c486418, c577cad), which would
have failed `.github/workflows/file-size-cap.yml` (cap=500) on next push.

Extracted the three per-path bodies to sibling sub-modules under
`phase_continue_prefills/`:

  - `run_standard.rs`         (215 LoC) — `run_standard_chunk_loop`:
    single-stream chunked body (mixed_forward + EP broadcast paths).
  - `run_batched_prefill.rs`  (126 LoC) — `run_batched_prefill_step`:
    Q12 N-stream batched-prefill step (active.is_empty case).
  - `run_batched_mixed.rs`    (151 LoC) — `run_batched_mixed_step`:
    Q12 Phase 5 batched mixed (decode+prefill) step.

Parent `phase_continue_prefills.rs` shrinks to 276 LoC — the
dispatcher only (continue_in_progress_prefills + gating comments).
Each sub-module is ≤250 LoC per `CLAUDE.md` core directive #4.

`pub(super)` on the helper fns; `mod ...` declarations + per-fn `use`
import in the parent. Dispatch behavior is unchanged.

`cargo build`, `cargo test -p spark-server --lib`, and
`cargo clippy --workspace --all-targets` all green.
