+++ crates/spark-server/src/scheduler/phase_continue_prefills.rs
@@ -8,16 +8,36 @@
 //!
 //! Returns `did_mixed_step` so the caller can skip the standalone decode
 //! call (mixed forward already processed decode logits).
+//!
+//! Layout: this file is the dispatcher only; the three per-path bodies
+//! live in the sibling sub-modules under `phase_continue_prefills/` to
+//! keep each unit ≤250 LoC per `crates/.../CLAUDE.md` core directive #4
+//! and ≤500 LoC per `.github/workflows/file-size-cap.yml`.
+//!
+//!  - `run_standard`        — single-stream chunked-prefill body
+//!                            (mixed_forward or plain prefill_chunk).
+//!  - `run_batched_prefill` — Q12 N-stream batched-prefill step.
+//!  - `run_batched_mixed`   — Q12 Phase 5 batched mixed (decode+prefill) step.
 
-use anyhow::Result;
-use spark_model::traits::{Model, PrefillSlice, SequenceState};
-use spark_runtime::gpu::DevicePtr;
+#[path = "phase_continue_prefills/run_batched_mixed.rs"]
+mod run_batched_mixed;
+#[path = "phase_continue_prefills/run_batched_prefill.rs"]
+mod run_batched_prefill;
+#[path = "phase_continue_prefills/run_standard.rs"]
+mod run_standard;
+
 use std::time::Instant;
 
-use super::decode_logits_step::process_decode_logits;
+use spark_model::traits::Model;
+
 use super::phase_promote_prefills::promote_completed_prefills;
-use super::*;
+use super::sample_token;
+use super::types::{ActiveSeq, PrefillInProgress};
 use crate::scheduling_policy::{ActiveSeqTiming, SchedulingPolicy};
+
+use run_batched_mixed::run_batched_mixed_step;
+use run_batched_prefill::run_batched_prefill_step;
+use run_standard::run_standard_chunk_loop;
 
 #[allow(clippy::too_many_arguments)]
 pub(super) fn continue_in_progress_prefills(
@@ -66,7 +86,7 @@ pub(super) fn continue_in_progress_prefills(
     // asymmetric 24+131 s TTFT documented in qwen-refactor notes §6.
     //
     // Phase 4a (commit 2ff926d): active.is_empty() case → prefill_batch_chunk.
-    // Phase 5 (this commit): active.is_nonempty() case → mixed_forward_batch
+    // Phase 5 (commit a542463): active.is_nonempty() case → mixed_forward_batch
     // (N decode tokens + M prefill chunks fused). Lifts the implicit MTP /
     // self-spec / N-gram-spec gating since those only apply to active
     // sequences, not freshly-prefilling ones — the active-side decode is
@@ -253,464 +273,4 @@ pub(super) fn continue_in_progress_prefills(
     );
 
     did_mixed_step
-}
-
-/// Inner loop: try mixed_forward first when conditions allow; else fall
-/// back to plain prefill_chunk + EP broadcast.
-#[allow(clippy::too_many_arguments)]
-fn run_standard_chunk_loop(
-    model: &dyn Model,
-    p: &mut PrefillInProgress,
-    idx: usize,
-    active: &mut Vec<ActiveSeq>,
-    max_prefill_tokens: usize,
-    prefill_stream: u64,
-    prefill_event: u64,
-    use_mtp: bool,
-    use_self_speculative: bool,
-    use_ngram_speculative: bool,
-    think_end_token: Option<u32>,
-    think_start_token: Option<u32>,
-    tool_call_start_token: Option<u32>,
-    tool_call_end_token: Option<u32>,
-    reflection_suppress_ids: &[u32],
-    adaptive_sampling: bool,
-    completed_indices: &mut Vec<(usize, Option<u32>)>,
-    did_mixed_step: &mut bool,
-) {
-    // Single chunk per call — the outer scheduler loop re-enters this
-    // function on the very next iteration to advance the next stream
-    // or the next chunk. This yield keeps fairness across pending
-    // requests (Q12: avoids back-to-back chunked prefill monopolising
-    // the scheduler).
-    let remaining = p.prompt_tokens.len() - p.chunk_offset;
-    // MLA correctness gate: Atlas has no `prefill_attention_paged_mla_*`
-    // kernel; the existing MLA prefill at qwen3_attention/prefill.rs:1723
-    // only attends over the current chunk's K/V, so multi-chunk prefill
-    // silently corrupts attention output. Force single-chunk until a
-    // paged-MLA prefill kernel lands. Hurts cold TTFT on long MLA
-    // prompts but preserves correctness.
-    let effective_max = if model.is_mla() {
-        remaining
-    } else {
-        max_prefill_tokens
-    };
-    let mut chunk_len = remaining.min(effective_max);
-    let is_last = p.chunk_offset + chunk_len >= p.prompt_tokens.len();
-    // Align intermediate chunks to GDN WY4 boundary (4 tokens).
-    if !is_last && chunk_len >= 4 {
-        chunk_len = (chunk_len / 4) * 4;
-    }
-
-    // ── Mixed forward: fuse prefill chunk + decode in one pass ──
-    // ATLAS_BISECT_NO_MIX=1 forces this branch to false so we can
-    // diagnose whether the chunked-prefill+concurrent CUDA-700 lives
-    // inside `mixed_forward` (active+prefill fused) vs the pure
-    // decode-batch path.
-    let no_mix_bisect = std::env::var("ATLAS_BISECT_NO_MIX")
-        .map(|v| v == "1" || v.to_lowercase() == "true")
-        .unwrap_or(false);
-    let can_mix = !no_mix_bisect
-        && !active.is_empty()
-        && !model.is_ep()
-        && !use_mtp
-        && !use_self_speculative
-        && !use_ngram_speculative;
-
-    if can_mix {
-        let decode_tokens: Vec<u32> = active.iter().map(|a| a.last_token).collect();
-        let mut decode_refs: Vec<&mut SequenceState> =
-            active.iter_mut().map(|a| &mut a.seq).collect();
-        let t0_mixed = Instant::now();
-
-        match model.mixed_forward(
-            &decode_tokens,
-            &mut decode_refs,
-            &p.prompt_tokens,
-            &mut p.seq,
-            p.chunk_offset,
-            chunk_len,
-            is_last,
-            prefill_stream,
-        ) {
-            Ok(result) => {
-                p.chunk_offset += chunk_len;
-                tracing::info!(
-                    "Mixed forward: prefill {}/{} tokens + {} decode",
-                    p.chunk_offset,
-                    p.prompt_tokens.len(),
-                    decode_tokens.len(),
-                );
-
-                // Process prefill logits (if last chunk).
-                if is_last {
-                    if let Err(e) = model.normalize_ssm_states(&p.seq, prefill_stream) {
-                        tracing::warn!("SSM state normalization failed: {e:#}");
-                    }
-                    let _ = model.record_event(prefill_event, prefill_stream);
-                    let _ = model.stream_wait_event(model.default_stream(), prefill_event);
-                    match sample_token(
-                        model,
-                        result.prefill_logits,
-                        p.temperature,
-                        p.top_k,
-                        p.top_p,
-                        &p.eos_tokens,
-                    ) {
-                        Ok(first) => {
-                            tracing::info!("Mixed prefill first token: {first}");
-                            completed_indices.push((idx, Some(first)));
-                        }
-                        Err(e) => {
-                            tracing::error!("Mixed prefill sampling: {e:#}");
-                            completed_indices.push((idx, None));
-                        }
-                    }
-                }
-
-                // Process decode logits for active sequences.
-                let _ = model.record_event(prefill_event, prefill_stream);
-                let _ = model.stream_wait_event(model.default_stream(), prefill_event);
-                process_decode_logits(
-                    model,
-                    active,
-                    result.decode_logits,
-                    t0_mixed,
-                    think_end_token,
-                    think_start_token,
-                    tool_call_start_token,
-                    tool_call_end_token,
-                    reflection_suppress_ids,
-                    adaptive_sampling,
-                );
-                *did_mixed_step = true;
-            }
-            Err(e) => {
-                tracing::error!("Mixed forward error: {e:#}");
-                completed_indices.push((idx, None));
-            }
-        }
-        return;
-    }
-
-    // ── Standard path: prefill chunk only, decode separately ──
-    // EP: broadcast chunk tokens to worker (bulk, single NCCL op).
-    let ep_ok = (|| -> Result<()> {
-        model.ep_broadcast_cmd(0xFFFFFFF0)?;
-        model.ep_broadcast_cmd(chunk_len as u32)?;
-        model.ep_broadcast_cmd(p.chunk_offset as u32)?;
-        model.ep_broadcast_cmd(p.prompt_tokens.len() as u32)?;
-        model.ep_broadcast_tokens(&p.prompt_tokens)?;
-        Ok(())
-    })();
-    if let Err(e) = ep_ok {
-        tracing::error!("EP broadcast chunk: {e:#}");
-        completed_indices.push((idx, None));
-        return;
-    }
-
-    match model.prefill_chunk(
-        &p.prompt_tokens,
-        &mut p.seq,
-        p.chunk_offset,
-        chunk_len,
-        is_last,
-        prefill_stream,
-    ) {
-        Ok(logits) => {
-            p.chunk_offset += chunk_len;
-            tracing::info!(
-                "Prefill chunk {}/{} tokens",
-                p.chunk_offset,
-                p.prompt_tokens.len(),
-            );
-            // Normalize SSM states after EVERY chunk to prevent state drift.
-            if let Err(e) = model.normalize_ssm_states(&p.seq, prefill_stream) {
-                tracing::warn!("SSM state normalization failed: {e:#}");
-            }
-            if is_last {
-                let _ = model.record_event(prefill_event, prefill_stream);
-                let _ = model.stream_wait_event(model.default_stream(), prefill_event);
-                match sample_token(
-                    model,
-                    logits,
-                    p.temperature,
-                    p.top_k,
-                    p.top_p,
-                    &p.eos_tokens,
-                ) {
-                    Ok(first) => {
-                        tracing::info!("Prefill first token: {first}");
-                        completed_indices.push((idx, Some(first)));
-                    }
-                    Err(e) => {
-                        tracing::error!("Chunked prefill argmax: {e:#}");
-                        completed_indices.push((idx, None));
-                    }
-                }
-            }
-        }
-        Err(e) => {
-            tracing::error!("Prefill chunk error: {e:#}");
-            completed_indices.push((idx, None));
-        }
-    }
-}
-
-/// Batched-prefill step (Q12): advance every prefilling stream by one chunk
-/// in a single `model.prefill_batch_chunk` call. Records first-token sample
-/// in `completed_indices` for any stream that just finished its last chunk.
-///
-/// Phase 4a (default-impl wiring): the model's default `prefill_batch_chunk`
-/// loops over single-stream `prefill_chunk`. No kernel batching yet — the
-/// behavioural win is fairness (every stream advances per iteration vs the
-/// FIFO `prefilling.first_mut()` starvation). Phase 2/3 replace the default
-/// impl with batched kernel dispatch for true L2-amortised throughput.
-fn run_batched_prefill_step(
-    model: &dyn Model,
-    prefilling: &mut [PrefillInProgress],
-    completed_indices: &mut Vec<(usize, Option<u32>)>,
-    max_prefill_tokens: usize,
-    prefill_stream: u64,
-    prefill_event: u64,
-) {
-    // Build per-stream chunk_len (capped at max_prefill_tokens) and
-    // is_last_chunk flag, then construct PrefillSlice borrowing each
-    // stream's prompt_tokens and seq.
-    //
-    // Capture per-stream chunk_len up-front so we can advance
-    // `chunk_offset` after the model call (the slices borrow `&mut p.seq`
-    // but not `&mut p.chunk_offset`, so post-call mutation is permitted
-    // once the slices vec is dropped).
-    let n = prefilling.len();
-    let mut chunk_lens: Vec<usize> = Vec::with_capacity(n);
-    let mut is_last_flags: Vec<bool> = Vec::with_capacity(n);
-    for p in prefilling.iter() {
-        let remaining = p.prompt_tokens.len() - p.chunk_offset;
-        // Same MLA correctness gate as `run_standard_chunk_loop` — MLA
-        // models lack a paged-MLA prefill kernel so multi-chunk prefill
-        // silently corrupts attention. Force single-chunk for MLA.
-        let effective_max = if model.is_mla() { remaining } else { max_prefill_tokens };
-        let mut chunk_len = remaining.min(effective_max);
-        let is_last = p.chunk_offset + chunk_len >= p.prompt_tokens.len();
-        // Align intermediate chunks to GDN WY4 boundary (4 tokens).
-        if !is_last && chunk_len >= 4 {
-            chunk_len = (chunk_len / 4) * 4;
-        }
-        chunk_lens.push(chunk_len);
-        is_last_flags.push(is_last);
-    }
-
-    // Build PrefillSlice borrows. Each slice borrows `&p.prompt_tokens`
-    // (immutable) and `&mut p.seq` from a distinct `&mut PrefillInProgress`,
-    // which is sound because the fields are disjoint.
-    let mut slices: Vec<PrefillSlice<'_>> = prefilling
-        .iter_mut()
-        .enumerate()
-        .map(|(i, p)| PrefillSlice {
-            prompt_tokens: &p.prompt_tokens,
-            seq: &mut p.seq,
-            chunk_start: p.chunk_offset,
-            chunk_len: chunk_lens[i],
-            is_last_chunk: is_last_flags[i],
-        })
-        .collect();
-
-    let t0_batch = Instant::now();
-    let logits_per_stream = match model.prefill_batch_chunk(&mut slices, prefill_stream) {
-        Ok(v) => v,
-        Err(e) => {
-            tracing::error!("Batched prefill error (streams={n}): {e:#}");
-            // Mark every stream as failed so they get freed in
-            // `promote_completed_prefills`.
-            for i in 0..n {
-                completed_indices.push((i, None));
-            }
-            return;
-        }
-    };
-    drop(slices); // release the &mut p.seq borrows so we can advance chunk_offset
-
-    // Sync prefill stream → default stream so subsequent decode sees
-    // the prefill writes. Mirrors the existing single-stream path.
-    let _ = model.record_event(prefill_event, prefill_stream);
-    let _ = model.stream_wait_event(model.default_stream(), prefill_event);
-
-    debug_assert_eq!(logits_per_stream.len(), n, "prefill_batch_chunk returned wrong logit count");
-
-    // Advance offsets and sample first token where the chunk just completed.
-    for (i, p) in prefilling.iter_mut().enumerate() {
-        p.chunk_offset += chunk_lens[i];
-        if !is_last_flags[i] {
-            continue;
-        }
-        let logits = logits_per_stream[i];
-        if logits == DevicePtr::NULL {
-            tracing::error!(
-                "Batched prefill: stream {i} marked is_last but model returned NULL logits",
-            );
-            completed_indices.push((i, None));
-            continue;
-        }
-        match sample_token(model, logits, p.temperature, p.top_k, p.top_p, &p.eos_tokens) {
-            Ok(first) => {
-                tracing::info!(
-                    "Batched prefill[{i}/{n}] first token: {first} (chunk_len={}, total_tokens={})",
-                    chunk_lens[i],
-                    p.prompt_tokens.len(),
-                );
-                completed_indices.push((i, Some(first)));
-            }
-            Err(e) => {
-                tracing::error!("Batched prefill[{i}] sampling: {e:#}");
-                completed_indices.push((i, None));
-            }
-        }
-    }
-
-    let elapsed = t0_batch.elapsed().as_micros();
-    if elapsed > 1000 {
-        tracing::debug!("Batched prefill step: {n} streams, {elapsed}µs total");
-    }
-}
-
-/// Batched mixed step (Q12, Phase 5): N decode tokens + M prefill chunks
-/// fused via `model.mixed_forward_batch`. Like `run_batched_prefill_step`
-/// but for the active-nonempty case — replaces the existing single-prefill
-/// `mixed_forward` call with the M-stream variant.
-///
-/// On success sets `did_mixed_step = true` so the caller skips the standalone
-/// decode dispatch (active sequences' next tokens are already sampled here).
-///
-/// Phase 5 uses the default trait impl which serializes (`decode_batch` then
-/// per-stream `prefill_chunk` loop). Phase 2/3 will override with true
-/// kernel-level batched mixed forward.
-#[allow(clippy::too_many_arguments)]
-fn run_batched_mixed_step(
-    model: &dyn Model,
-    active: &mut Vec<ActiveSeq>,
-    prefilling: &mut [PrefillInProgress],
-    completed_indices: &mut Vec<(usize, Option<u32>)>,
-    max_prefill_tokens: usize,
-    prefill_stream: u64,
-    prefill_event: u64,
-    t0_step: Instant,
-    think_end_token: Option<u32>,
-    think_start_token: Option<u32>,
-    tool_call_start_token: Option<u32>,
-    tool_call_end_token: Option<u32>,
-    reflection_suppress_ids: &[u32],
-    adaptive_sampling: bool,
-    did_mixed_step: &mut bool,
-) {
-    let n_prefill = prefilling.len();
-    let n_decode = active.len();
-
-    // Capture per-stream chunk_len + is_last (same MLA gate + WY4 alignment
-    // as `run_batched_prefill_step`).
-    let mut chunk_lens: Vec<usize> = Vec::with_capacity(n_prefill);
-    let mut is_last_flags: Vec<bool> = Vec::with_capacity(n_prefill);
-    for p in prefilling.iter() {
-        let remaining = p.prompt_tokens.len() - p.chunk_offset;
-        let effective_max = if model.is_mla() { remaining } else { max_prefill_tokens };
-        let mut chunk_len = remaining.min(effective_max);
-        let is_last = p.chunk_offset + chunk_len >= p.prompt_tokens.len();
-        if !is_last && chunk_len >= 4 {
-            chunk_len = (chunk_len / 4) * 4;
-        }
-        chunk_lens.push(chunk_len);
-        is_last_flags.push(is_last);
-    }
-
-    // Gather decode-side inputs.
-    let decode_tokens: Vec<u32> = active.iter().map(|a| a.last_token).collect();
-
-    // Build slices in a temporary scope so the &mut borrows on prefilling
-    // and active drop before we re-borrow active mutably for
-    // `process_decode_logits`.
-    let result = {
-        let mut decode_refs: Vec<&mut SequenceState> =
-            active.iter_mut().map(|a| &mut a.seq).collect();
-        let mut prefill_slices: Vec<PrefillSlice<'_>> = prefilling
-            .iter_mut()
-            .enumerate()
-            .map(|(i, p)| PrefillSlice {
-                prompt_tokens: &p.prompt_tokens,
-                seq: &mut p.seq,
-                chunk_start: p.chunk_offset,
-                chunk_len: chunk_lens[i],
-                is_last_chunk: is_last_flags[i],
-            })
-            .collect();
-        model.mixed_forward_batch(
-            &decode_tokens,
-            &mut decode_refs,
-            &mut prefill_slices,
-            prefill_stream,
-        )
-    };
-
-    let result = match result {
-        Ok(r) => r,
-        Err(e) => {
-            tracing::error!(
-                "Mixed-batch forward error (n_decode={n_decode}, n_prefill={n_prefill}): {e:#}",
-            );
-            for i in 0..n_prefill {
-                completed_indices.push((i, None));
-            }
-            return;
-        }
-    };
-
-    let _ = model.record_event(prefill_event, prefill_stream);
-    let _ = model.stream_wait_event(model.default_stream(), prefill_event);
-
-    // Advance prefill offsets and sample first tokens for streams that just
-    // finished their last chunk.
-    debug_assert_eq!(result.prefill_logits.len(), n_prefill);
-    for (i, p) in prefilling.iter_mut().enumerate() {
-        p.chunk_offset += chunk_lens[i];
-        if !is_last_flags[i] {
-            continue;
-        }
-        let logits = result.prefill_logits[i];
-        if logits == DevicePtr::NULL {
-            tracing::error!("Mixed-batch: stream {i} marked is_last but model returned NULL logits");
-            completed_indices.push((i, None));
-            continue;
-        }
-        match sample_token(model, logits, p.temperature, p.top_k, p.top_p, &p.eos_tokens) {
-            Ok(first) => {
-                tracing::info!(
-                    "Mixed-batch prefill[{i}/{n_prefill}] first token: {first} (chunk_len={}, total_tokens={})",
-                    chunk_lens[i],
-                    p.prompt_tokens.len(),
-                );
-                completed_indices.push((i, Some(first)));
-            }
-            Err(e) => {
-                tracing::error!("Mixed-batch prefill[{i}] sampling: {e:#}");
-                completed_indices.push((i, None));
-            }
-        }
-    }
-
-    // Process decode logits for the active lanes — mirrors what
-    // `run_standard_chunk_loop`'s mixed_forward branch does.
-    if n_decode > 0 && result.decode_logits != DevicePtr::NULL {
-        process_decode_logits(
-            model,
-            active,
-            result.decode_logits,
-            t0_step,
-            think_end_token,
-            think_start_token,
-            tool_call_start_token,
-            tool_call_end_token,
-            reflection_suppress_ids,
-            adaptive_sampling,
-        );
-    }
-    *did_mixed_step = true;
 }
